// src/modules/court/enums/court-amenities.enum.ts
export enum CourtAmenities {
  WIFI = 'WIFI',
  LOCKER_ROOM = 'LOCKER_ROOM',
  PARKING = 'PARKING',
  BBQ_AREA = 'BBQ_AREA',
  BATHROOM = 'BATHROOM',
  LIGHTING = 'LIGHTING',
  COVERED = 'COVERED',
  BAR = 'BAR',
  RESTAURANT = 'RESTAURANT',
  WATER_FOUNTAIN = 'WATER_FOUNTAIN',
  BLEACHERS = 'BLEACHERS',
}